Bundle OpenStream shoud return ConcatenatedAssets if more than one asset
In debug mode, with the default configuration assets won't get
concatenated, which is a good thing. But in some cases (e.g.: using
Bundles.Url for downlading a script async), even in debug mode, if you
ask for a bundle url, it has to be concatenated. It has no sense to
serve the first asset of the bundle