Fix https://github.com/andrewdavey/cassette/issues/355
If a compiler, such as SASS or Less or Compass made a compile error previously the error would become "stuck" in the AppDomain and no further attempts to rebuild the bundle would be made until the AppDomain recycled. This changes the code to re-initialize the bundle once if an exception occurred during the initial rebuilding, which allows compile errors to be reevaluated without a rebuild.